moved custom table widget item to a separate class
